#2158 moving hook to wp files
