+++ src/frontend/src/apis/work-packages.api.ts
@@ -4,10 +4,11 @@
  */
 
 import axios from '../utils/axios';
-import { WbsNumber, WorkPackage, WorkPackageStage } from 'shared';
+import { WbsNumber, WorkPackage, WorkPackageStage, WorkPackageTemplate } from 'shared';
 import { wbsPipe } from '../utils/pipes';
 import { apiUrls } from '../utils/urls';
 import { workPackageTransformer } from './transformers/work-packages.transformers';
+import { workPackageTemplateTransformer } from '../../../backend/src/transformers/work-package-template.transformer';
 
 export interface WorkPackageApiInputs {
   name: string;
@@ -102,5 +103,15 @@ export const getManyWorkPackages = (wbsNums: WbsNumber[]) => {
 export const slackUpcomingDeadlines = (deadline: Date) => {
   return axios.post<{ message: string }>(apiUrls.workPackagesSlackUpcomingDeadlines(), {
     deadline
+  });
+};
+
+/**
+ * Gets all the workpackage templates from the database
+ * @returns gets all the workpackage templates
+ */
+export const getAllWorkPackageTemplates = () => {
+  return axios.get<WorkPackageTemplate[]>(apiUrls.workPackageTemplates(), {
+    transformResponse: (data) => JSON.parse(data).map(workPackageTemplateTransformer)
   });
 };
